requirements.txt
aloe==0.2.0
Django==3.0.4
djangorestframework==3.11.0
